Relocated utils package
