+++ code/tutorlib/gui/app/app.py
@@ -17,7 +17,7 @@ from tutorlib.gui.dialogs.progress import ProgressPopup
 from tutorlib.gui.dialogs.submissions import SubmissionsDialog
 from tutorlib.gui.editor.delegate import TutorEditorDelegate
 from tutorlib.gui.editor.editor_window import TutorEditor
-from tutorlib.gui.utils.decorators import skip_if_attr_none
+from tutorlib.utils.decorators import skip_if_attr_none
 import tutorlib.gui.utils.messagebox as tkmessagebox
 from tutorlib.gui.utils.threading import exec_sync
 from tutorlib.interface.interpreter import Interpreter
