+++ code/tutorlib/gui/editor/io_binding.py
@@ -20,7 +20,8 @@
 ## It is a modification of idlelib/IOBinding.py
 
 from idlelib import IOBinding
-import tutorlib.gui.utils.messagebox as tkmessagebox
+
+import tutorlib.utils.messagebox as tkmessagebox
 
 
 class TutorIOBinding(IOBinding.IOBinding):
