+++ code/tutorlib/online/session.py
@@ -34,7 +34,7 @@ import urllib.parse
 import urllib.request
 
 from tutorlib.gui.dialogs.login import LoginDialog
-import tutorlib.gui.utils.messagebox as tkmessagebox
+import tutorlib.utils.messagebox as tkmessagebox
 from tutorlib.online.exceptions import AuthError, BadResponse, RequestError
 from tutorlib.online.parser import FormParser, strip_header
